+++ example_project/config/urls.py
@@ -3,5 +3,7 @@ from django.urls import include, path
 
 urlpatterns = [
     path("admin/", admin.site.urls),
+    path("accounts/", include("django.contrib.auth.urls")),
+    path("app/", include("modular_brix.portal.urls")),
     path("org/", include("modular_brix.foundation.organizations.urls")),
 ]